refactoring, trying to create barrel-files

+++ build.ps1
@@ -1,2 +1,2 @@
-Remove-Item -Path build -Recurse
+if (Test-Path build) { Remove-Item -Path build -Recurse }
 tsc
